Refactor, fix statistics output
