+++ metron-analytics/metron-statistics/src/test/java/org/apache/metron/statistics/approximation/HyperLogLogPlusFunctionsIntegrationTest.java
@@ -1,3 +1,20 @@
+/**
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership.  The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License.  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
 package org.apache.metron.statistics.approximation;
 
 import org.adrianwalker.multilinestring.Multiline;
@@ -11,6 +28,7 @@ import java.util.Map;
 import static org.hamcrest.CoreMatchers.equalTo;
 
 public class HyperLogLogPlusFunctionsIntegrationTest {
+
   private Map<String, Object> values = new HashMap<String, Object>() {{
     put("val1", "mike");
     put("val2", "miklavcic");
@@ -78,4 +96,5 @@ public class HyperLogLogPlusFunctionsIntegrationTest {
     Long estimate = (Long) StellarProcessorUtils.run(hllpMergeRule, values);
     Assert.assertThat("Incorrect cardinality returned", estimate, equalTo(4L));
   }
+
 }
